projeto_api/pyproject.toml
[project]
name = "projeto-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "pydantic[email]>=2.11.7",
    "sqlmodel>=0.0.24",
    "strawberry-graphql[fastapi]>=0.281.0",
    "uvicorn>=0.35.0",
]
